GameManager now caches full object names
If you use an abbreviated name, it will do a Find the first time, and then store the full object name in the fullnames dictionary to later reference, since Find is not the most effectient for deep hierarchies.  i.e. you don't need to say /Canvas/SelectionPanel/SelectionPanel/Phone  You can just say Phone and it will store the full path the first time it encounters it.